Fix first instantiation of Configuration and ApplicationManager
The first instantiation of Configuration and ApplicationManager were previsouly at different places in
the code and where made the first time we had to use the objects. It's now fixed by instantiating them
for the first time in the SmartphoneFrame constuctor, which is the first method called in the Main.main().
This solution provides better understanding of the state of the Configuration and the ApplicationManager
when their objects are instantiated.