Fix run() method becomes open() in ApplicationManager
The run() method was used to open an application, that is to say two distinct operations (start it the first time it is clicked,
then display it if it had already been started previously).
This change renames the run() method to open() to have a name that reflects better its behaviour.
Furthermore, a run() method is created to contain the operations of adding the app to the running app list and adding the app to
the contentPanel in order for it to be displayed.