Fix resources loading by using getResourceAsStream()
This changes introduces the use of the `yourClass.class.getResourceAsStream()` method to retrieve any image / icon / resource file, which works both with pre-compiled files and in a `jar` archive.
The resources loading was previously made with `new File()` or `getResource()` alone (that is to say without `ImageIO.read()`), which can't be used to load resources in a `jar` archive.