Fix infinite loop when the 'config.json' file is not found
This change fixes the infinite loop that could occur if the config.json file was not found or could not be read.
This loop was caused by a recursive call to the getInstance() method if an exception was thrown in this same getInstance() method.
This change causes the getInstance() method to return null if an exception is thrown in it.