Fix lockbutton timer when turning off the phone
Before this change the timer was started not only on the `mousePressed` event but also on the `mouseClicked`event.
This change stops the timer when the `mouseReleased` is fired and thus removes the timer to start on a `mouseClicked` event.
This change removes therefore the hack of clicking multiple times on the lockbutton during 4 seconds to turn off the phone.