Fix showConfirmationDialog asking multiple times when deleting album folder
Before this fix, when the deleteAlbum button was clicked,
the JOptionPane.showConfirmationDialog() was called as many times as there were pictures
in the album. This behaviour was happening because the addPictures() method (called when we want to import
pictures via the JFileChooser) was internally calling the initEvents() method each time a picture was added.
The result of this was that the deleteAlbum button had as many listeners as there were pictures added and thus
shows the JOptionPane.showConfirmationDialog() as many time as there where listeners.

This fix replaces the initEvents() call in the addPictures() method by a call to a new method
'updateAlbumView(AlbumModel)' that refreshes the whole GUI by creating a complete new albumView.