Add WrapLayout class to make a grid-like view of thumbnails
It's not possible to use a FlowLayout combined with a JScrollPane to build a grid-like view because the FlowLayout doesn't change its preferredSize dynamically, and therefore the JScrollPane never displays its scrollbars.
This issue is fixed by making a WrapLayout, extending FlowLayout, that does update dynamically its preferredSize.
The JScrollPane will therefore work as expected.